requirements.txt
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extensions==0.7.1
uritemplate==4.1.1
PyYAML==6.0
django-filter==22.1
sentry-sdk==1.9.10